requirements.txt
certifi>= 14.05.14
python_dateutil>= 2.5.3
setuptools
urllib3

requests